// src/modules/chat/chat.interface.ts
export enum ChatType {
  PRIVATE = 'private',
  GROUP = 'group',
}

export enum ChatGroupType {
  ERROR = 'error',
  ADMIN = 'admin',
  SUPERADMIN = 'superadmin',
}

export interface Chat {
  id: number;
  name: string;
  telegramChatId: number;
  type: ChatType;
  chatGroupType: ChatGroupType | null;
  verified: boolean;

  createdAt: Date;
  updatedAt: Date;
}

export interface TelegramChat {
  id: number;
  name: string;
  type: ChatType;
}

export interface CreateChatDto extends Partial<Omit<Chat, 'id' | 'createdAt' | 'updatedAt'>> {
  name: string;
  telegramChatId: number;
  type: ChatType;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface UpdateChatDto extends Partial<Omit<Chat, 'id' | 'createdAt' | 'updatedAt'>> {}